myproject/requirements.txt
psycopg2 == 2.8.4
django-cors-headers == 3.2.0